Fixing critical flaw in CSS Uploader design.
The CSS Uploader and CSS Kit work together by keeping skin directories for
CSS, images, and JavaScript at one level; like this:

skin/
|- images/
|- css/
`- javascript/

The problem is that we reorganize this into the following structure:

share/nlw/
|- images/
|  `- skin/
|- css/
|  `- skin/
`- javascript/
   `- skin/

The flaw doesn't recognize that url() resources identified in a stylesheet are
relative to the URI of the CSS file. This is what people are writing in their
stylesheets for the CSS Uploader, or by using the CSS Kit:

  background: url(../images/spacer.gif);

This translates to the following URI:

  $CSS_URI/../images/spacer.gif

That fails because we put the images at the following URI:

  $CSS_URI/../../images/skin/spacer.gif

This will cause the resource to be unfound (404), and breaks the skin.

This quick RewriteRule resolves that issue by mapping "right-side-in" skin
directory layout into the current physical layout on the server.

Please release soon.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@5652 593ce9bc-2028-0410-bb98-c566b643a948