Add rudimentary support for DELETE on /data/workspaces/:ws/users/:username.
For Story: Expose "remove member" from Workspace via REST

If the acting user is an admin remove target_user from the
target_workspace. If the acting_user and target_user are the same,
also remove target_user from the workspace. In either of those
cases if the user is not present in the workspace give a 404[1].

Otherwise give a 401 or 403 on DELETE, 405 on any other method.

Since only DELETE is made operational in this change, there are
no content-type issues to deal with.

t/live/rest/workspace-user.t could be less complex, but I chose
to use the REST API, rather than the Perl API, for most of the
user adding and other associated diddling not direcdtly 
related to the testing of DELETE. I've made an effort to make
the intent of the code clear, with comments and such.

In the story I interpreted "admin" to mean "workspace admin".
Other options are "business admin" and "technical admin". 
What should their powers be in this situation?

Next:

* confirm this is in the ballpark
* deal with TODO
* resolve technical/business admin question
* document (presumably in st-rest-docs?)

[1] In the case of target_user and acting_user being equal, it is
not clear if a 404 or 403 should be returned when the acting_user
is not in the target_workspace. The natural flow of the code is 
for it to be a 404. Presumably it is okay for a user to know 
whether or not they are in a workspace. I've marked the test 
associated with this particular case with a TODO. Input welcome.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7349 593ce9bc-2028-0410-bb98-c566b643a948