Implement symmetry between adding and removing users to a workspace. This means that when doing either a
* business admin
* technical admin
* or admin of the workspace

Can do both. A user can also remove themselves from a 
workspace.

This was mostly a matter of changing the gate that 
causes an error to be sent when the user doesn't match
the right permissions. I did however, change that from
a 401 to a 403. A 401 is only sent when we want the 
request to attempt authentication. At this stage we want
to say, "you ain't going to get it, not you, so don't
bother trying again"

The vast majority of the code is addition of tests
to cover the various paths.

Meanwhile, in the background, there's some ongoing 
debate about the meaning of business_admin which 
may require some changes to this and WorkspaceUser.pm.
Stay tuned.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7362 593ce9bc-2028-0410-bb98-c566b643a948