Implement setting the permission set of a workspace in the REST API.
The original story for this suggested a new location in 
the REST API for handling this change. However the suggested
form broke the existing grammar/patterns of the REST API 
so Shawn and I talked about it and came up with an alternate
way.

the permission set of a workspace can be set by a JSON
PUT to the workspace. The key 'permission_set', if set
to a valid Socialtext::Workspace::Permissions::PermissionSet,
causes the workspace to be updated with the value.

This action can be done by the same roles that can DELETE
a workspace, so adjusted the code accordingly.

I also added a permission_set key to the outgoing workspace
information given on a GET request (for admins).

Will put docs for this in st-rest-docs tomorrow.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7434 593ce9bc-2028-0410-bb98-c566b643a948