Test Case Tags is adjusted to work for socialcalc
I'm not happy with this for several reasons:
A) The save button sometimes takes extrodinarily love for new pages - blank
pages even.  On occasion, the save button takes over 60 seconds and causes and
error.  This may be a performance issue -- or it may be a wikitests issue.

B) The Save Tag button doesn't seem to save when clicked - it consistently 
hangs the 3rd time it's clicked.  So, I click the button twice - with 
pauses between - for every page. I hate that because is is not the real user
experience.

---> This still needs work, but it passes the tests, and I'm off to bed.
Search needs my attention worse anyway.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7516 593ce9bc-2028-0410-bb98-c566b643a948