Adjusted the test case to reflect the spreadsheets in test-data
Als, {bz 614} is an issue.  Basically, any spreadsheet will be found when you
search for "1.5" because the version of socialcalc is 1.5.  I wrote the test
with a knowledge of the bug so if the bug is fixed the test will fail - this
is by design; it is sideways TDD.

This may be an issue if tests create spreadsheets and don't clean up after
themselves, or if the version of socialcalc changes, or if the bug is fixed.
If and when those occur, I will change the tests



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7520 593ce9bc-2028-0410-bb98-c566b643a948