Fixes synchonicity issues with the test case
The test case I checked in last night used an awkward and unrealistic series
of simulated button clicks to add tags.  This is a problem because if the
real series of button clicks were suddenly broken but the simulator still
worked, we'd never know.

I refactored the test to be like test_case_tagging, which uses a more
realistic set of clicks.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7521 593ce9bc-2028-0410-bb98-c566b643a948