Remove traces of the "password" attribute in the LDAP attr_map.
We don't use it for anything, so why do we still have it kicking around?

Also added a note to make it clear that we should leave in the fix for Bug
{bz 621}, until such time as a migration script is put in place; we _may_ have
customers out there with this attribute configured and we still want to make
sure that we don't pull their passwords from the LDAP directory by accident.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7626 593ce9bc-2028-0410-bb98-c566b643a948