Fixed the test to work under Internet Explorer When running under IE, some of the text looks different to selenium because som eof the whitespace is stripped.
So matching "foo bar" does not work.  I had to convert that to qr/foo\s*bar/

I hated doing this, as it made the test ugly, but the alternative is to
basically ignore feedback from the test under IE because "oh yeah, that test
is broken under IE" - thus, when the software really does break, we won't
catch it.

Broken windows lead to graffiti, and, eventually, theft and destruction


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7709 593ce9bc-2028-0410-bb98-c566b643a948