Lots of changes to calctests.
I started by trying to modify the existing, "good" calctests to work
regardless of the installation of people or dashboard.  That went rather well.

Then I converted ExportToHTML to a calc test.

Then I took a long hard look at what was let, and a lot of it was setting
workspace permissions, or dealing with RSS, or specific tests for the editor
that did not apply to socialcalc.  So I removed them from OSR_tescases,
deleted them from the wiki, and removed those test cases from SVN.

As a result, cacltests are really close to done.  I should finish shortly.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7742 593ce9bc-2028-0410-bb98-c566b643a948