Enhance the web interface timings logging somewhat, but not all the way to the spec:
* instead of listing the explicit URL, log the path from
  url_map that was matched
* log the :ws, :pname etc parameters
* log the query string
* stick the HTTP status code in there

The goal (via swd) here is to log successful requests so 
that unsuccessful requests do not muddy the timings data.
I made an easy mistake to make with the way the code works
and assumed I could use the HTTP status code as the signal
of success, as we are at the HTTP layer of the code (the
apache handler).

Unfortunately, success is ill defined, and the code frequently
does not set the http status header at all. So what we have
now is that we log timings when:

* the status code is between 200 and 399 (success and redirection)
* the status header is not set at all

S2 does not set the status code at all, except when it knows
it needs to set a redirection. Mike and the REST API do most
of the time, but not all the time (see my recent commit that
adds some status headers).

Similarly, how authentication and authorization is handled is
not centralized enough to be able to perform a clean trap and
declare "oh lookie, we're doing auth, let's not do a log here".

I'm not sure I see an easy clean way out of this. What I've done
for now is not log when we know the status code to indicate failure
and to include in the log information that shows which paths don't
set status.

In the future it might be useful to twist things in the code around
a bit such that there is one single location through which 

a. All requests flow (sort of have that now).
2) Exceptions are thrown at the point of redirects and authentication
   requests and there is one single place where they are caught
   and handled, or if not a single place, then they are at least
   re-raised.

This latter makes it possible to throw an HTTP401 somewhere in the
depths of the code with the expectation that it will be effectively
handled somewhere, by something, but that something is the choice
of the upstream code. This would make Challenger handling quite a
bit more flexible, and would make for more readable code.

Some of this exist in various forms but it is not unified.

Python's WSGI presents a good model for this sort of thing.

Anyway: the logging: what's there allows for a bit of data
gathering that could lead to further investigation.

WEB,GET,/:WS/INDEX.CGI,UNDEF,q:slartibartfast,ws:admin,\
  [handler_run:0.301174,web_auth:0.011318,overall:0.314714]

Please let me know where it is out of alignment with expectations.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7744 593ce9bc-2028-0410-bb98-c566b643a948