checkpointing changes to Timer code, for sake of review.
The basic goal is to have global access to some timers
so we can easily throughout the code add in timing information
and then report on it near some exit point.

It's assumed that since there are no process concurrency issues
with ST, we can use a bona fide global as a sort of singleton-esque
thingie.

Thus, make $Timings in Socialtext::Timer. At the entry point,
call Socialtext::Timer->Reset(). This clears $Timings to {}
and sets (in $Timings) an 'overall' timer. Further calls
to Socialtext::Timer->Start($name) creates a $name timer.
Stop($name) will stop the timer and change the value of key
$name, in $Timings, to the elapsed time.

When Socialtext::Timer->Report() is called, any timer that
is not stopped, is StopPED and a hash of timername => elapsed
time pairs is returned.

The test in t/Socialtext/Timer.t should make it pretty clear.
If this maps to what people are hoping for, I'll change the
existing timings logging to use this new model.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7799 593ce9bc-2028-0410-bb98-c566b643a948