Using the new Socialtext::Timer->{Reset,Start,Stop,Report} API, log reports from the Socialtext::CLI entry point. Because of the way the ceqlotron works this will also log for each command the ceqlotron runs (the specific commands, not the from_input stuff, if that's even happening anymore).
The Timer API currently assumes there is on global Timings
hash on which information is built, so you can have one for
each entry point, starting multiple timers of different
names, but having only one overall and one Report.

Shawn and I have discussed a variety of ways to make more 
complicated things happen, but at the moment, this is sort
of the next step in the iterating. I'll commit some more
::Timer oriented that refactors the ACTION and WEB log
in a bit.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7815 593ce9bc-2028-0410-bb98-c566b643a948