Add a timer for HTML generation when displaying a page (for the initial page display) in S2.
Check this in just as an example of some of the ways in which
the timers can be used, and some of the code changes that
are necessary. In this case a call was moved outside a
giant hash so the timer could be stopped and started.

It's important to know/consider that variables assignement
in Perl, especially of long strings, can be a bit expensive,
so this change to assign to $page_html is a performance
deoptimization for the sake of gathering information.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7819 593ce9bc-2028-0410-bb98-c566b643a948