This should fix {bz: 691}, where the new $page->preview_text returns a raw text (unesscaped) and that's used directly as element attributes. Mostly something like this:
    $preview = $page->preview_text;
    return qq{<a title="$preview" ...>};

Specifically, it'll break when preview text contains double quotes, which makes
the generated HTML looks something like this:

    <a title="It is a shark with "laser" on its head" href="index.cgi?shark">shark</a>

Which is interpreted as something like this on most browsers:

    <a title="It is a shark with "
       laser=""
       on=""
       its=""
       head=""
       href="index.cgi?shark">shark</a>

It can be easily seen how it breaks the semantics. If there's pair of unescaped <> 
in preview_text, like this:

    <a title="Harry potter cast "the spell"<acrabadabra>" href="#">Harry</a>

It's parsed like this:

    <a title="Harry potter cast " the="" spell="">
        <acrabadabra>" href="#">Harry</acrabadabra>
    </a>

Notice that there's a new element <acrabadabra> with text node: 

    " href="#">Harry'
    
inside.  This part will be the text displayed. And that's how it breaks
the display of everywhere. In page display and in recent changes,
and dashboards... anywhere where link is previewed.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@7890 593ce9bc-2028-0410-bb98-c566b643a948