Running the tests pointed out some undefined warnings leading to exposing a bug in the Timer stuff that I'm not sure why the tests didn't point out, but...?
tv_interval wants an array of a particular form, that returned
by gettimeofday in list context. But I just willy nilly 
stuck a '-offset' in there which blew that all to bit so.

So now use time() (the one from Time::HiRes) and don't 
bother with gettimeofday and tv_interval at all.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8035 593ce9bc-2028-0410-bb98-c566b643a948