The new Wikitext::Socialtext::Parser dies if it can't parse something.
We use this module to generate previews for mouseovers.

This fix will catch a preview generation error and return an empty
preview in the rare case that a page is not parsable. The error is
written to the log so we can see them and fix the parser when it
needs a fix.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8037 593ce9bc-2028-0410-bb98-c566b643a948