Quick hack to prevent an application error when using Socialtext::Timer in S::TT2::Renderer::render().
It turns out, much to my surprise, that in some circumstances render()
gets re-entered by itself. This happens when dealing with user preferences.

template/element/settings/preferences_settings_section has this line:

    [% preference.form_element %]

form_element is a sub on S::PreferencesPlugin that _sometimes_ eventually
calls S::TT2::Renderer::render().

I spoke with Shawn, and we debated around about how to deal with this.
Initially we were thinking this should be fixed as there's probably 
other places where the same thing happens.

However, as I poked around with trying to fix it, any solution I could
come up with wasn't quite going through to being complete and it was
destroying the very lightweight nature of the Timer code (requiring
more complex data structures).

So for the time being I've made it so if pause gets called twice,
the second causes the recorded time to be 'reentered' and just bails.

Obviously this is a stopgap for now, but I wanted to get something in there
that allowed settings and preferences to work again.

Suggestions welcome.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8045 593ce9bc-2028-0410-bb98-c566b643a948