Port the guanxi schema into NLW
There is some pretty complex stuff going on here.  We need to conditionally create and update the tables, because they exist on beta boxes but not on an official release of the code.  We use a new migration (25) to update the storage table and skip the socialtext-2-to-3.sql schema migration on these beta boxes.

From the migration README:

    This migration exists to migrate beta branches of code back to a known point.

    There are two issues in some released beta code.  This migration corrects the
    released beta code:

    1) The released beta code didn't have a user_id column in the 'storage' table.
    2) The released beta code already has event and people tables 

    All of these tables are added by the /etc/socialtext/db/socialtext-2-to-3.sql
    file, and this migration will bump the Socialtext Schema version to skip this
    sql update from happening.

    This migration will only run if the Schema version is 2; all boxes that we
    know of that had released beta code are at this version.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8053 593ce9bc-2028-0410-bb98-c566b643a948