Make the Timer code reentrant, thus making it possible for a timer to continue itself and pause itself multiple times while only tracking the interval between the first continue and the last pause.
In actual implementation what it does is ignore any pause that
is not properly balanced by a continue, there's a counter that
increments on every start, and decrements on every stop.

There are probably other ways to do this, but this way preserved
the what seems to me rather clean and simple API of the timers.

I suspect it is not entirely immunue to confusion. For example
under some error conditions a counter may stay high when it shouldn't.
In my opinion, this is an indicative of a problem in the code as it
is in the timers, so leaving it go seems okay.

In other words, when analyzing the data, it makes sense to:

* throw away outliers when doing statistical analysis
* examine outliers when trying to figure out what's weird about
  the system

Please review this code. It passes the tests and doesn't cause
the bug Ken found, but...


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8061 593ce9bc-2028-0410-bb98-c566b643a948