When daemonizing, make sure we close *all* filehandles first.
This does cause problems in that `__DATA__` gets closed, so we need to make
sure that we grab that data _before_ we daemonize ourselves.

`slapd` also refuses to start if we close STDIN/STDOUT before reopening them,
so we don't bother closing them outright but just reopen them to `/dev/null`.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8092 593ce9bc-2028-0410-bb98-c566b643a948