Fix {bz 506}, named web link causes bad link formatting
The basic problem here was that the regular expression was 
being greedy, so a " early in a line could pair up with the
second " in the label of a wafl phrase, making for a really
long label.

This patch changes the quote matching to "[^"]" which 
seems to get the job done and not blow up any Page*.t or 
Formatter/*.t tests.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8260 593ce9bc-2028-0410-bb98-c566b643a948