Another attempt at getting "_user_root()" cleaned up, this time basing it off of "_user_checkout_dir()".
Part of the issue here is that lookups based on __FILE__ or %INC may include
relative paths.  This becomes a problem we we've done a chdir out of the NLW
directory and then call something that works its way down here into
ST::AppConfig.  e.g. when we're importing workspaces with LDAP enabled.

In these cases, we want to make sure that "_user_root()" returns a path
relative to the actual directory we started in, *not* the directory that we're
in now.

Passes unit tests on Topaz, but haven't seen what it'll do on an appliance
yet.  Will know more after testrunner has a run through on this...


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8465 593ce9bc-2028-0410-bb98-c566b643a948