When importing workspaces, match up users by "username", not "email_address".  Fixes {bz 761}.
If we try to match them up by "email address" *and* we've accidentally
exported a "Deleted User", they're just about guaranteed to *not* match up
with anything in the user factories.  Unfortunately, though, their *username*
might actually be in use (as they may in fact still be a real/active user).

A scenario where this could occur easily?  Exporting a workspace from an LDAP
enabled appliance, when the LDAP server is down.  During the export, none of
the user records will look up correctly against the LDAP directory, and
*everyone* will be exported as a Deleted User.

When we look them up by username, however, we'll find them again in the user
factories, and eliminate potential username conflicts.  _If_ it turns out that
we couldn't find them in the user factories, we should have no troubles
actually creating a user record for them (as we know that we have *no*
username conflict). :)


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8466 593ce9bc-2028-0410-bb98-c566b643a948