Log everything to "~/.nlw/log/nlw.log", if you're running in a dev-env.
Got sick and tired of not being able to see what was being logged when running
things in a dev-env.  I'd written up a 0.5pt Story for this back in March, but
its never made it into an iteration.

Finally decided to JFDI as I'm wasting far more than 0.5pts trying to figure
out whats going on in my dev-env.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8508 593ce9bc-2028-0410-bb98-c566b643a948