Put our PID file in `/tmp/` instead of $ENV{HOME}, so that it works on appliances when run as "www-data".
Ken had mentioned that he had problems with `st-bootstrap-openldap` when
running it on an appliance during wikitests, partly due its trying to drop the
PID file in the home directory for the user (which in this case isn't writable
by the user).

So... stick the PID file off in `/tmp/` instead, and add the UID to the
filename so that we don't get clashes between users.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8730 593ce9bc-2028-0410-bb98-c566b643a948