One of the technical details with the new photo upload requires that the photos be stored in the DBMS. In order to do so, two things need to happen: 1) the DB user needs to be a super user for the system, and 2) we need to update our schema.
The schema update is handled in the standard way.

The DB user update happens during a migration that also
looks at existing users and imports photos ( from URL's ) into the 
DBMS with these newly created privs.

To round things out, on new installations we need to make sure that
the DB user is created with superuser privs.

Note: DB user in this case is separate and distinct from system user.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8935 593ce9bc-2028-0410-bb98-c566b643a948