If HTTP requests to Technorati time out, skip the tests.
Otherwise, go ahead and run them.

If we find that this turns out to not be an adequate/suitable test for "is
Technorati available", repeated failures will help us to figure out either
what sort of test we should have for "is it available", or give us enough info
to be able to make the tests more robust so they don't fail (retry perhaps?).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@8995 593ce9bc-2028-0410-bb98-c566b643a948