Re-fix detection of "build" dir for "out of branch" testing.
We auto-detect the "build" dir so we can use it later on to find the template
root.  Move some of the path mangling out of when we try to assemble the
template root directory over to where we detect the build dir.  This lets us
simplify the "assemble the template root dir" bits so that they're less likely
to break in the future.

Yes, we've had multiple kicks at this can, and we've re-broken it again
repeatedly with subsequent commits.  This version should (hopefully) be a bit
more resistant to breakage.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9002 593ce9bc-2028-0410-bb98-c566b643a948