Implement the first failing visual test for {bz: 855}
The test requires S3 theme and browser need to login to admin workspace
first.

The test content is testing if "Export" and "All Pages | All Tags |
All Files" are overlapping with eachother. In the description
of {bz: 855} it happens when browser window is narrower. In this case
I(gugod) found that 800px is _narrow_ enough to trigger this bug.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9043 593ce9bc-2028-0410-bb98-c566b643a948