Fix broken fdefs; don't set up re-indexing for a dev-env.
If we're actually trying to bootstrap a dev-env from scratch, we won't be able
to run `st-admin list-workspaces` and the bootstrap fails.

I traced this back to entering the codestream in r8980 on the 08/08 iteration
branch, which then landed in trunk in r9032.  If you've already got a dev-env
in place and are re-creating it, you'll never hit this as a bug.  If you're
bootstrapping from scratch, though, it'd halt you dead in your tracks.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9092 593ce9bc-2028-0410-bb98-c566b643a948