Cleanup leftover workspaces at the end of a unit test.
If any workspaces were created during the run of a unit test and were left
behind, complain to the user and then go clean them out (giving us a slightly
saner state for the next test suite to run in).

Reduces the amount of cruft that gets carried over at the end of a test run,
similarly to how we're already cleaning out User records that are left behind.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9134 593ce9bc-2028-0410-bb98-c566b643a948