Migration for first/last names from person table.
Moves first/last names set in the person table into the "User" table, but only
wherever the "User" table entry is empty or just whitespace.

Removes the first_name, last_name, name and email columns.

middle_name was also removed in this iteration, so the 8-to-9 migration should
still be valid.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9163 593ce9bc-2028-0410-bb98-c566b643a948