Allow test fixtures to conflict with other fixtures.
Until `T::ST::Environment` is updated to make use of this, we're still going
to do "clean things out on every test".  Being able to detect conflicts
between the requested fixture and an already bootstrapped fixture, though,
brings us closer to being able to re-use existing fixtures and not have to
flush everything out and recreated it for every single test suite.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9173 593ce9bc-2028-0410-bb98-c566b643a948