Basic sanity checking for "did we parse a valid line of CSV?"
If we parse more than 'x' fields (where 'x' is "the number of fields that
exist for required/mandatory data), then treat the data as "good" and pass it
on down the line to try to add/update the user based on that data.

Unit test updates include a line of "non-CSV", and a line of "tab-separated
CSV", followed by a line of good CSV (which should still continue to be
processed and the user added/updated from that line).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9233 593ce9bc-2028-0410-bb98-c566b643a948