ZOMG, what a day!
It took us a very long time to realize a passing test was passing incorrectly and we've been testing widgets incorrectly all along.

Now we know how to wait properly for a widget to finish loading before checking it.

ALong the way we made a lot of improvement to Test.Visual.

Testing is good. I feel like it has dragged us through the stuff we need to know to be truly effective.




git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9472 593ce9bc-2028-0410-bb98-c566b643a948