Factor out the "Edit/Comment" buttons so they can be re-used on the bottom of the page.  Fixes {bz 888}
Refactored buttons accept an "id_suffix" parameter that gets tacked onto the
end of the HTML element ID, which is used here to help give different IDs to
the buttons when they're on the "top" or "bottom".

JS updated so that it now attaches the appropriate actions to the buttons when
they appear on the bottom of the page too.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9577 593ce9bc-2028-0410-bb98-c566b643a948