Make unit tests run *MUCH* faster, by caching fixtures from test to test.
Currently requires that you run with NLW_LIVE_DANGEROUSLY=1; some of the tests
still have hidden or poorly defined expectations of how things should look at
the start of the test (or how badly they mangle the data in the test).

But, if you're running your own unit test repeatedly while doing dev, this can
help speed up your testing substantially by not forcably cleaning out the test
environment each time it starts your test.

If any conflicts are found between the fixtures requested and those in the
cache, though, the "clean" fixture is prepended to the list of fixtures
automatically so that things start off with a clean slate.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@9981 593ce9bc-2028-0410-bb98-c566b643a948