Fixes {bz: 1293}.
Althought it appears that jQuery does not select the iframe html element,
but evidently, doing $("html").css("overflow", "hidden") also reflow
the HTML in wysiwyg editor iframe.

Prevents the html/iframe reflow by detecting if it's required to hide
window scrollbar. When it is edit mode, usually we already adjust the
editor height such that scrollbar is already gone.

This solution is better then detecting if it's editor mode because
that would make lightbox code depend on wikiwyg code.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10099 593ce9bc-2028-0410-bb98-c566b643a948