I don't see how this ever worked - fixed now.
The real watchlist function creates users, then clicks on order by user edited
to get a list.  This version never creates those users.  So the users are all
wikitester, so the order is determined by spelling.  I changed the
expectations to work.

Were we just not running calctests as part of the regular process?  I can't
see why this would work before.



git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10109 593ce9bc-2028-0410-bb98-c566b643a948