Use an in-memory cache for all User lookups.
What actually gets cached is the homonculus.  Default and LDAP homonculuses are cached separately.

A global variabled, $Socialtext::User::Cache::Enabled controls the User cache
behaviour.  We turn this on in the Apache PerlInitHandler, which is run before
each request.  By default, it is off.  

So, a side-effect of this change is that in scripts, "Default"-factory User
objects are no longer cached (but the caching behaviour can be easily
restored).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10122 593ce9bc-2028-0410-bb98-c566b643a948