Add test coverage for RecentChangesPlugin.
lando++ and I reverse-engineered these tests from the
existing behaviour of this plugin.

We hope that this covering this module, as well as the other
ST::Query::Plugin modules, we can eventually do a fearless
Refactor of the underlying architecture -- for example, do
sorting/paging in the DB where it belongs.  Before we can do
the "results fully from the db" Refactoring, we'll need the
LDAP user cache story done.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10410 593ce9bc-2028-0410-bb98-c566b643a948