Fixing {bz: 1372}.
Anice refactoring of code that handles mouse events on widgets.
This new approach uses event delegation, the memory consumtion is lower,
good for a page width many widgets.

A "controlselect" event handler is bound and canceled to get rid of
those small resize rectangles on IE. It's a IE specific event.

At this moemnt , there is still a chance to make the resize rectangles
show up when there are multiple widgets in the page. More fixes on coming.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10415 593ce9bc-2028-0410-bb98-c566b643a948