Refactor the `$user->delete()` functionality down into `ST::User::Base`.
When deleting _any_ type of user record, we want to go in and clean up any
entry that they might have in the "user_detail" table.  At the moment this is
an empty DELETE for LDAP users, but once we get their data cached in this
table it'll be more useful.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10595 593ce9bc-2028-0410-bb98-c566b643a948