Fix a widget editing bug found on Firefox 2, reproducible with these steps:
1. Edit any page, switch to Rich Text mode
2. Insert a new spreadsheet widget, give it a random title.
3. Click on the newly inserted widget

On Firefox 2, you'll see that the content of lightbox is doubled and
those save/cancel buttons are not working.

The real cause is that Firefox 2 fires "mouseup" event twice when clicking
on a widget and our widget_editing flag variable is decreasing too much,
such that the reentrant-preventing if statement in getWidgetInput() method
fails to work properly.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10606 593ce9bc-2028-0410-bb98-c566b643a948