Merge the "User" and "UserId" tables together into a single "users" table.
This is needed to implement LDAP-user caching (but we don't implement
any of that in this patch).

This helps clean up some of the ugliness that lived under the hood in
the way that users are handled.  Most notably, it helps clear up the
confusion surrounding "the user_id value", "a UserId object", "the
system unique id for a user", and "the driver_unique_id value".

User factories are now responsible for calling
`ST::User::Base->NewUserId()` to get a new unique user id when they
are creating a new user record in the database.

User factories are also now totally responsible for placing their user
records into the "users" table when they are created.  The Default
factory did this already, and this behaviour has been pushed down into
the LDAP factory as well.  Rather than relying on
`ST::UserId->create_if_necessary()`, the user factory is now
responsible for creating the equivalent record in the "users" table.
This may get refactored into a base class when we implement LDAP-user
caching.

The `ST::UserId` class has been removed, as has its test suite.  The
`t/Socialtext/User-deleted.t` test suite has also been removed as
there is now *no* reason that the "User" and "UserId" tables should
ever get out of sync with one another (as its now just *one table*).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10622 593ce9bc-2028-0410-bb98-c566b643a948