Add caching to LDAP Users.
Move Homunculus retrieval out of the DB from the Default Factory into the Base
homunculus.  This allows the LDAP Factory to use the same code.

The LDAP Factory will check db-retrieved homunculi against a fixed (for now)
TTL of 300 seconds.  If the db-retrieved homunculus is not fresh, the Factory
checks the LDAP server.

Whenever the LDAP Factory fetches a homunculus from the LDAP server, it caches
this entry in the database.

Calling ST::User->new(driver_unique_id => $foo) doesn't seem to work, but the
only thing that needs this currently is LTPA (outside of trunk).  We will
address this problem in a subsequent commit.

git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10644 593ce9bc-2028-0410-bb98-c566b643a948