Lazy-load the LDAP connection when looking up Users, plus other fixes.
This should bring [Story: Cache LDAP User Data] even closer to completion.

The LDAP factory will now wait until the last possible moment to connect to
the LDAP server, allowing it to check the long-term cache without overhead.
This also allows us to return cached copies in the face of connectivity
issues.

A number of tests needed to be updated to explicitly check the LDAP connection.

Expired LDAP users (i.e. their 'cached_at' is beyond the TTL) will get
re-checked on the LDAP server.  If the user is missing on the server, we now
return a ST::User::Deleted user object.  If we couldn't connect to the LDAP
server for whatever reason, just return the cached copy even though it's
expired.

Users with invalid driver_keys in the database are now treated like Deleted
users.

The fix for looking up users based on `driver_unique_id` is also in this
patch.  This was needed for LTPA integration.

"password" is no-longer a valid Search key when searching and LDAP store for
users.

Workspace-import.t is an odd test where the original "bad behaviour" that the
test was checking for got fixed as a result of this work.  I modified the test
to force the user to be Deleted so that the rest of it (the actual import
part) is still exercised.

There was much Refactoring and the Homunculi did rejoice.

Sorry for yet another mega-patch, folks. :(

gtermars++ for all of the guidance and pairing.

~stash


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10659 593ce9bc-2028-0410-bb98-c566b643a948