Change the way we check for "did the user get refreshed?" and "who did the refresh?"
I know... "sleep" isn't the most elegant way to do it, but its simple and also
makes it obvious when/where the refresh of the user record got done.

This update also sets some bogus data into the DB prior to the refresh, so we
*know* that we not only updated the "cached_at" for the user data but that we
also actually went out and grabbed that info from LDAP again and populated the
DB with it.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10937 593ce9bc-2028-0410-bb98-c566b643a948