Migration script to prime the DB cache for LDAP users.
NOTE: if this migration fails to run or (for whatever reason) can't contact
LDAP in order to run through without error, *that's OK*.  No harm, no foul;
we're just trying to prime the DB cache with information on all of the known
LDAP users, _if_ we can.  If we can't, we'll have to wait for the user data to
get cached as part of the natural operation of the system; not a big deal.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10939 593ce9bc-2028-0410-bb98-c566b643a948