Leave a breadcrumb trail around, to help figure out why users aren't Authenticating.
After seeing this pop up in a few Support issues recently, where LDAP users
were unable to login because they had outstanding email confirmations, having
this info logged is going to be useful in the future.  If nothing else, its
going to make it a whole lot easier to figure out *why* these users are having
trouble logging in.


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@10958 593ce9bc-2028-0410-bb98-c566b643a948